Add option for downloading documentation

+++ _plugins/fetch_docs.rb
@@ -1,8 +1,8 @@
 #!/usr/bin/env ruby
 
-require 'yaml'
-require 'fileutils'
-require 'optimist'
+require "yaml"
+require "fileutils"
+require "optimist"
 
 opts = Optimist::options do
   version "fetch_docs 0.0.2 (c) The Infinispan team"
@@ -41,22 +41,22 @@ def get_docs(tmp, name, repo, branch, loc, docroot, docbase, verbose, attr_heade
 
   docs_src = "#{git_clone_path}/#{loc}"
   d = Dir.open docs_src
-  d.each {|f| FileUtils.cp_r("#{docs_src}/#{f}", target, :verbose => verbose) if f != "." and f != ".." and f != "Guardfile"}
+  d.each { |f| FileUtils.cp_r("#{docs_src}/#{f}", target, :verbose => verbose) if f != "." and f != ".." and f != "Guardfile" }
 
-  if attr_header != "" then
+  if attr_header != ""
     puts "Prepending attribute header to .adoc files:\n#{attr_header}"
-    Dir.glob("#{target}/**/*.adoc").each {|f|
+    Dir.glob("#{target}/**/*.adoc").each { |f|
       basedir = File.dirname(f)
       chapter = File.basename(f).start_with?("chapter")
       incbase = basedir[basedir.index(docbase)..-1]
       file_new = File.open("#{f}.new", "w")
       file_new.puts("#{attr_header}\n")
       file_old = File.open(f, "r")
-      file_old.each {|line|
-        if chapter then
+      file_old.each { |line|
+        if chapter
           file_new.puts(line)
         else
-          file_new.puts(line.sub( /include::/, "include::#{incbase}/"))
+          file_new.puts(line.sub(/include::/, "include::#{incbase}/"))
         end
       }
       file_old.close()
@@ -80,10 +80,10 @@ def get_maven_docs(htmlArtifact, pdfArtifact, docroot, docbase, docalias)
   FileUtils.mkdir_p target
 
   extract_maven_artifact(htmlArtifact, target)
-  if pdfArtifact != nil then
+  if pdfArtifact != nil
     extract_maven_artifact(pdfArtifact, target)
   end
-  if docalias != nil then
+  if docalias != nil
     aliastarget = File.expand_path("#{docroot}/#{docalias}")
     puts "    Alias #{docalias}"
     FileUtils.rm_rf aliastarget
@@ -91,42 +91,46 @@ def get_maven_docs(htmlArtifact, pdfArtifact, docroot, docbase, docalias)
   end
 end
 
-FileUtils.rm_rf(Dir.glob("docs/*"))
-Dir.mkdir("docs/") unless File.exists?("docs/")
-versions_xml_file = File.open("docs/versions.xml", "w")
-versions_xml_file.puts("<?xml version=\"1.0\" encoding=\"UTF-8\"?>");
-versions_xml_file.puts("<versions>");
-cfg["docs"].each do |type, tcfg|
-  puts "Processing #{type}"
-  tcfg.each do |ver, vcfg|
-    puts "  #{ver}"
+forceDocumentationDownload = (ENV.fetch("FORCE_DOCUMENTATION_DOWNLOAD") { "true" }).upcase
+if forceDocumentationDownload == "FALSE" and File.exists?("docs/versions.xml")
+  puts "Documentation exists. Skip the forced download..."
+else
+  FileUtils.rm_rf(Dir.glob("docs/*"))
+  Dir.mkdir("docs/") unless File.exists?("docs/")
+  versions_xml_file = File.open("docs/versions.xml", "w")
+  versions_xml_file.puts("<?xml version=\"1.0\" encoding=\"UTF-8\"?>")
+  versions_xml_file.puts("<versions>")
+  cfg["docs"].each do |type, tcfg|
+    puts "Processing #{type}"
+    tcfg.each do |ver, vcfg|
+      puts "  #{ver}"
 
-    attr_header = ""
-    asciidocattr = vcfg["asciidocattr"]
-    if asciidocattr != nil then
-      asciidocattr.each do |attr_name, attr_value|
-        attr_header = "#{attr_header}:#{attr_name}: #{attr_value.to_s}\n"
+      attr_header = ""
+      asciidocattr = vcfg["asciidocattr"]
+      if asciidocattr != nil
+        asciidocattr.each do |attr_name, attr_value|
+          attr_header = "#{attr_header}:#{attr_name}: #{attr_value.to_s}\n"
+        end
+      end
+
+      if type == "infinispan"
+        core = vcfg["core"]
+        valias = core["alias"]
+        get_maven_docs(core["html"], core["pdf"], "docs", "#{ver}", valias)
+        vname = if valias != nil then "#{ver} (#{valias})" else "#{ver}" end
+        versions_xml_file.puts("<version name=\"#{vname}\" path=\"#{ver}\" />")
       end
     end
-
-    if type == "infinispan"
-      core = vcfg["core"]
-      valias = core["alias"]
-      get_maven_docs(core["html"], core["pdf"], "docs", "#{ver}", valias)
-      vname = if valias != nil then "#{ver} (#{valias})" else "#{ver}" end
-      versions_xml_file.puts("<version name=\"#{vname}\" path=\"#{ver}\" />");
-    end
   end
-end
 
-# cpp and dotnet clients need the same special treatment
-["cpp","dotnet"].each do |client|
-  cfg["hotrod"][client]["downloads"].each do |label, artifact|
-    if artifact.include?("docs_version")
+  # cpp and dotnet clients need the same special treatment
+  ["cpp", "dotnet"].each do |client|
+    cfg["hotrod"][client]["downloads"].each do |label, artifact|
+      if artifact.include?("docs_version")
         docs_version = artifact["docs_version"]
         url = artifact["download"]
         fullname = File.basename(url)
-        name = File.basename(url,".zip")
+        name = File.basename(url, ".zip")
         %x( wget #{url} -O _tmp.zip)
         %x( unzip _tmp.zip "*-Source/documentation/*" -d _tmp)
         %x( mkdir -p docs/hotrod-clients/#{client}/docs )
@@ -136,50 +140,51 @@ end
           %x( asciidoctor #{f} )
         end
         %x( rm -rf _tmp _tmp.zip )
+      end
     end
   end
-end
 
-# and then it's operator's turn
-cfg["ispn_operator"].each do |version, sub|
-  puts "#{version} wget"
-  zipUrl=sub["zip-url"]
-  puts "#{version} wget #{zipUrl}"
-  %x( wget #{zipUrl} -O _optmp.zip)
-  %x( unzip _optmp.zip "*documentation/*" -d _optmp)
-  Dir.glob("_optmp/**/*.asciidoc").each do |f|
-    %x( asciidoctor #{f} )
+  # and then it's operator's turn
+  cfg["ispn_operator"].each do |version, sub|
+    puts "#{version} wget"
+    zipUrl = sub["zip-url"]
+    puts "#{version} wget #{zipUrl}"
+    %x( wget #{zipUrl} -O _optmp.zip)
+    %x( unzip _optmp.zip "*documentation/*" -d _optmp)
+    Dir.glob("_optmp/**/*.asciidoc").each do |f|
+      %x( asciidoctor #{f} )
+    end
+    Dir.glob("_optmp/**/*.html").each do |f|
+      %x( cp -r #{f} _optmp )
+    end
+    %x( mkdir -p docs/infinispan-operator/#{version}/ )
+    %x( mkdir -p docs/infinispan-operator/topics/images/ )
+    %x( mv _optmp/*.html "docs/infinispan-operator/#{version}/" )
+    #Use images only on master. Causes harmless "cannot stat" messages.
+    %x( cp -r _optmp/infinispan-operator-master/documentation/asciidoc/topics/images/* "docs/infinispan-operator/topics/images/" )
+    %x( rm -rf _optmp* )
   end
-  Dir.glob("_optmp/**/*.html").each do |f|
-    %x( cp -r #{f} _optmp )
-  end
-  %x( mkdir -p docs/infinispan-operator/#{version}/ )
-  %x( mkdir -p docs/infinispan-operator/topics/images/ )
-  %x( mv _optmp/*.html "docs/infinispan-operator/#{version}/" )
-  #Use images only on master. Causes harmless "cannot stat" messages.
-  %x( cp -r _optmp/infinispan-operator-master/documentation/asciidoc/topics/images/* "docs/infinispan-operator/topics/images/" )
-  %x( rm -rf _optmp* )
-end
 
-# now for the spring boot starter docs
-cfg["sb_starter"].each do |version, sub|
-  puts "#{version} wget"
-  zipUrl=sub["zip-url"]
-  puts "#{version} wget #{zipUrl}"
-  %x( wget #{zipUrl} -O _sbtmp.zip)
-  %x( unzip _sbtmp.zip "*documentation/*" -d _sbtmp)
-  Dir.glob("_sbtmp/**/*.asciidoc").each do |f|
-    %x( asciidoctor #{f} )
+  # now for the spring boot starter docs
+  cfg["sb_starter"].each do |version, sub|
+    puts "#{version} wget"
+    zipUrl = sub["zip-url"]
+    puts "#{version} wget #{zipUrl}"
+    %x( wget #{zipUrl} -O _sbtmp.zip)
+    %x( unzip _sbtmp.zip "*documentation/*" -d _sbtmp)
+    Dir.glob("_sbtmp/**/*.asciidoc").each do |f|
+      %x( asciidoctor #{f} )
+    end
+    Dir.glob("_sbtmp/**/*.html").each do |f|
+      %x( cp -r #{f} _sbtmp )
+    end
+    %x( mkdir -p docs/infinispan-spring-boot/#{version}/ )
+    %x( mv _sbtmp/*.html "docs/infinispan-spring-boot/#{version}/" )
+    %x( rm -rf _sbtmp* )
   end
-  Dir.glob("_sbtmp/**/*.html").each do |f|
-    %x( cp -r #{f} _sbtmp )
-  end
-  %x( mkdir -p docs/infinispan-spring-boot/#{version}/ )
-  %x( mv _sbtmp/*.html "docs/infinispan-spring-boot/#{version}/" )
-  %x( rm -rf _sbtmp* )
-end
 
-versions_xml_file.puts("</versions>");
-versions_xml_file.close()
+  versions_xml_file.puts("</versions>")
+  versions_xml_file.close()
+end
 
 puts "Time elapsed #{Time.now - beginning} seconds"
